requirements.txt
python-telegram-bot[job-queue]==21.4
SQLAlchemy==2.0.25
alembic==1.13.2
asyncpg==0.29.0
python-dotenv==1.0.1
openpyxl==3.1.2
matplotlib==3.8.2
pandas==2.1.4
<<<<<<< HEAD
=======
python-dateutil==2.8.2
>>>>>>> origin/main
tzdata==2023.4
